Reorganise step models Move the step models into store/WorkflowStepStore and store/WorkflowStepStore/Step.
