+++ packages/lib-classifier/src/store/WorkflowStepStore/Step/index.js
@@ -1,0 +1,1 @@
+export { default } from './Step'
