+++ packages/lib-classifier/src/store/WorkflowStepStore/index.js
@@ -1,0 +1,1 @@
+export { default } from './WorkflowStepStore'
